chore(media-sync-worker): migrate cronjob to internal app
