--- .github/workflows/cronjob-docker.yml
@@ -1,34 +1,0 @@
-name: Build and Push Cronjob Image
-
-on:
-  push:
-    branches: [main]
-    paths:
-      - 'apps/cloud/cronjob/**'
-      - 'packages/ts-shared/**'
-      - 'packages/lark-utils/**'
-      - 'packages/pixiv-client/**'
-
-jobs:
-  build-and-push:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Login to Docker Hub
-        uses: docker/login-action@v3
-        with:
-          username: ${{ secrets.DOCKER_HUB_USERNAME }}
-          password: ${{ secrets.DOCKER_HUB_ACCESS_TOKEN }}
-
-      - name: Build and push
-        uses: docker/build-push-action@v6
-        with:
-          context: .
-          file: apps/cloud/cronjob/Dockerfile
-          push: true
-          tags: |
-            chiwiio/cronjob-work:latest
-            chiwiio/cronjob-work:${{ github.sha }}
-          build-args: |
-            GIT_SHA=${{ github.sha }}
