--- apps/cloud/cronjob/.env.example
@@ -1,22 +1,0 @@
-# MongoDB 配置
-MONGO_HOST=localhost
-MONGO_INITDB_ROOT_USERNAME=root
-MONGO_INITDB_ROOT_PASSWORD=your_secure_password_here
-
-# Redis 配置
-REDIS_HOST=localhost
-REDIS_PASSWORD=your_redis_password_here
-
-# 飞书开放平台配置
-APP_ID=cli_xxxxxxxxxxxxxxxxxxxx
-APP_SECRET=your_feishu_app_secret_here
-SELF_CHAT_ID=oc_xxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-
-# HTTP 服务认证
-HTTP_SECRET=your_random_secret_key_here
-
-# Bangumi API 访问令牌
-BANGUMI_ACCESS_TOKEN=your_bangumi_access_token_here
-
-# 可选：端口配置（如有需要）
-# PORT=3000
