+++ apps/media-sync-worker/.env.example
@@ -1,0 +1,26 @@
+# MongoDB 配置
+MONGO_HOST=localhost
+MONGO_PORT=27017
+MONGO_INITDB_ROOT_USERNAME=root
+MONGO_INITDB_ROOT_PASSWORD=your_secure_password_here
+MONGO_CONNECT_TIMEOUT_MS=10000
+
+# Redis 配置
+REDIS_HOST=localhost
+REDIS_PORT=6379
+REDIS_PASSWORD=your_redis_password_here
+
+# 飞书开放平台配置
+APP_ID=cli_xxxxxxxxxxxxxxxxxxxx
+APP_SECRET=your_feishu_app_secret_here
+SELF_CHAT_ID=oc_xxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+
+# HTTP 服务认证
+HTTP_SECRET=your_random_secret_key_here
+
+# Bangumi API 访问令牌
+BANGUMI_ACCESS_TOKEN=your_bangumi_access_token_here
+
+# 部署验证开关
+DISABLE_SCHEDULES=false
+DISABLE_CONSUMER=false
