+++ apps/media-sync-worker/.gitignore
