+++ apps/media-sync-worker/src/mongo/service.ts
