+++ apps/media-sync-worker/src/mongo/types.ts
