+++ apps/media-sync-worker/src/service/dailyDownload.ts
