+++ apps/media-sync-worker/tsconfig.json
