+++ package.json
@@ -6,7 +6,7 @@
     "apps/lark-proxy",
     "apps/monitor-dashboard",
     "apps/monitor-dashboard-web",
-    "apps/cloud/cronjob",
+    "apps/media-sync-worker",
     "packages/ts-shared",
     "packages/pixiv-client",
     "packages/lark-utils"
